fix: now core-embed should work again
