--- experimental/go-sdk/cmd/example/.gitignore
@@ -1,1 +1,0 @@
-embed-core.bin
