--- experimental/go-sdk/cmd/example/embed-core.bin
@@ -1,1 +1,0 @@
-# this is written by go generate
