+++ experimental/go-sdk/cmd/examples/embed/.gitignore
@@ -1,0 +1,1 @@
+embed-core.bin
